+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add `insert` and `extract` instructions.
 - Add missing `unwrap_*` and `is_*` functions to types. ([#35](https://github.com/fabianschuiki/llhd/issues/35))
+- Add struct and array type parsing.
 
 ### Changed
 - Rename `as_*` functions on types to `unwrap_*`.